requirements.txt
Python >= 3.6
numpy==1.21.5
pandas==1.4.2
scipy==1.7.3
matplotlib==3.5.1
scikit-learn==1.1.3
nltk==3.7
stemming==1.0.1
sentence-transformers==2.3.1
torch==2.2.0
transformers==4.37.2
gensim==4.1.2
hdbscan==0.8.28
umap-learn==0.5.5
